Moved gdt and tss into a more fiting 'cpu' directory
